+++ src/main/java/net/dungeon_difficulty/config/ClientConfig.java
@@ -1,0 +1,9 @@
+package net.dungeon_difficulty.config;
+
+import net.minecraft.util.Rarity;
+
+public class ClientConfig {
+    public boolean enable_overriding_enchantment_rarity = true;
+    public boolean enable_scaled_items_rarity = true;
+    public Rarity scaled_item_rarity_max = Rarity.RARE;
+}