+++ src/main/java/net/dungeon_difficulty/mixin/ItemStackMixin.java
@@ -1,11 +1,16 @@
 package net.dungeon_difficulty.mixin;
 
+import net.dungeon_difficulty.DungeonDifficulty;
+import net.dungeon_difficulty.logic.RarityHelper;
 import net.minecraft.entity.attribute.EntityAttributeModifier;
 import net.minecraft.item.ItemStack;
 import net.dungeon_difficulty.logic.ItemScaling;
+import net.minecraft.util.Rarity;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.Redirect;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import java.util.UUID;
 
@@ -23,5 +28,21 @@ public class ItemStackMixin {
             return ItemScaling.ItemAccessor.hardCodedAttackSpeedModifier();
         }
         return instance.getId();
+    }
+
+    private ItemStack itemStack() {
+        return (ItemStack) (Object) this;
+    }
+
+    @Inject(method = "getRarity", at = @At("RETURN"), cancellable = true)
+    private void injected(CallbackInfoReturnable<Rarity> cir) {
+        var value = cir.getReturnValue();
+        var nbt = itemStack().getNbt();
+        if (nbt != null && nbt.contains(ItemScaling.ALREADY_SCALED_NBT_KEY)
+                && DungeonDifficulty.clientConfig.value.enable_scaled_items_rarity
+                && value.ordinal() <= DungeonDifficulty.clientConfig.value.scaled_item_rarity_max.ordinal()) {
+            var newValue = RarityHelper.increasedRarity(value, 1);
+            cir.setReturnValue(newValue);
+        }
     }
 }